requirements.txt
Babel==2.8.0
cryptography==2.6.1
flake8==3.7.9
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
html2text==2020.1.16
html5lib==1.0.1
importlib-metadata==0.23
itsdangerous==0.24
jdcal==1.4.1
jedi==0.13.2
Jinja2==2.10
lxml==4.3.2
matplotlib==3.0.2
neovim==0.3.1
ninja2==0.1
numpy==1.16.2
oauthlib==2.1.0
pandas==1.0.3
psycopg2==2.8.4
pycrypto==2.6.1
pydot==1.4.1
pyflakes==2.1.1
pynvim==0.3.2
PyPDF2==1.26.0
pypi==2.1
python-dotenv==0.14.0
python-editor==1.0.4
python-openid==2.2.5
request==2019.4.13
requests==2.21.0
requests-html==0.10.0
requests-oauthlib==1.0.0
responses==0.9.0
rsa==4.6
scrubadub==1.2.1
seaborn==0.10.1
simplegeneric==0.8.1
simplejson==3.16.0
spidev==3.4
SQLAlchemy==1.3.11
virtualenv==20.0.26
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
